backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
aiosqlite>=0.20.0
groq>=0.11.0
httpx>=0.27.0
python-multipart>=0.0.9
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
langchain>=0.2.0
langgraph>=0.2.0
pypdf>=4.0.0
python-docx>=1.1.0
